add プロジェクトのステート管理のためコンテキストを別ファイルに移動
